feat: complete some APIS in Restaurant and Dish feature and adjust somes in Drone feature

+++ src/main/java/com/fast_food_drone_delivery_system/dto/request/DroneUpdateRequest.java
@@ -5,13 +5,15 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 public class DroneUpdateRequest {
     private String status; // should be one of DroneStatus values
-    private Double currentLat;
-    private Double currentLng;
-    private Double batteryLevel;
+    private BigDecimal currentLat;
+    private BigDecimal currentLng;
+    private BigDecimal batteryLevel;
 }
